#3 Add compareScreenshot.js command and assertion
This commit adds a custom [command and assertion](https://gist.github.com/umarmw/cb06569dfb50dfab3a5e) which compares two images.  It also adds the baseline images that will be used when comparing screenshots. The mismatch tolerance was set to 6% which means that only images that are at minimum 94% similar will be accepted.