#1 Dockerize app
This commit is an attempt to dockerize the whole app. At the moment it can only run the tests in chrome, but this will be fixed in future commits. The Dockerfile in this commit makes an image that can make containers which include node/npm/nightwatch. The docker-compose.yml in this commit attempts to simplify the process to build and run the tests.